refactor: 양방향 관계 삭제
일정-댓글 양방향 관계를 댓글 -> 일정 단방향 관계로 수정
cascade 적용해 부모 삭제 시 자식도 삭제 되도록 설정

+++ src/main/java/com/example/tododevelop/service/UserService.java
@@ -55,6 +55,7 @@ public class UserService {
     // 전체 회원 조회
     public AllUserResponseDto findAllUser() {
         List<UserEntity> findAllUser = userRepository.findAll();
+        //
         List<UserResponseDto> usersDto = findAllUser.stream()
                 .map(UserResponseDto::new)
                 .collect(Collectors.toList());
@@ -68,15 +69,21 @@ public class UserService {
     }
 
     // 유저 정보 수정
-    public UserResponseDto modifyUserInfo(Long id, UserModifyRequestDto dto) {
+    public UserResponseDto modifyUserInfo(Long id, Long userId, UserModifyRequestDto dto) {
         UserEntity findUser = findByIdOrElseThrow(id);
+        if (findUser.getId() != userId) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "본인만 삭제 가능합니다.");
+        }
         findUser.modifyUserInfo(dto);
         return new UserResponseDto(findUser);
     }
 
     // 유저 삭제
-    public void deleteUser(long id) {
+    public void deleteUser(long id, Long userId) {
         UserEntity findUser = findByIdOrElseThrow(id);
+        if (findUser.getId() != userId) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "본인만 삭제 가능합니다.");
+        }
         userRepository.delete(findUser);
     }
